Export SvelteComponentDev as API
