+++ src/runtime/ambient.ts
@@ -1,4 +1,3 @@
 declare module '*.svelte' {
-	// eslint-disable-next-line import/no-unresolved
-	export { default } from 'svelte/internal/ComponentApi';
+	export { SvelteComponentDev as default } from 'svelte/internal';
 }
