+++ src/runtime/internal/dev.ts
@@ -79,8 +79,24 @@ export function set_data_dev(text, data) {
 	text.data = data;
 }
 
+
+type Props = Record<string, any>;
+export interface SvelteComponentDev {
+    $set(props?: Props): void;
+    $on<T = any>(event: string, callback: (event: CustomEvent<T>) => void): () => void;
+    $destroy(): void;
+    [accessor: string]: any;
+}
+
 export class SvelteComponentDev extends SvelteComponent {
-	constructor(options) {
+	constructor(options: {
+        target: Element;
+        anchor?: Element;
+        props?: Props;
+        hydrate?: boolean;
+		intro?: boolean;
+		$$inline?: boolean;
+    }) {
 		if (!options || (!options.target && !options.$$inline)) {
 			throw new Error(`'target' is a required option`);
 		}
