+++ src/runtime/internal/index.ts
@@ -12,5 +12,4 @@ export * from './ssr';
 export * from './transitions';
 export * from './utils';
 export * from './Component';
-export * from './ComponentApi';
 export * from './dev';
